+++ tests/New-DbaDbEncryptionKey.Tests.ps1
@@ -40,7 +40,9 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
         if ($delmastercert) {
             $mastercert | Remove-DbaDbCertificate
         }
-        $db | Remove-DbaDatabase
+        if ($db) {
+            $db | Remove-DbaDatabase
+        }
     }
 
     Context "Command actually works" {
